Doing experiments with matter cad as mesh
